# api/openapi.yml
openapi: '3.0.3'
info:
  title: 'Zone42 API'
  description: 'Zone42 administrator api'
  version: '1.0'
servers:
  - url: https://api.zone-42.com
security:
  - bearerAuth: []
paths:
  /zones:
    get:
      summary: 'list of all zones'
      parameters:
        - $ref: '#/components/parameters/start'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/ascending'
        - $ref: '#/components/parameters/q'
      responses:
        200:
          $ref: '#/components/responses/list_response'
        400:
          $ref: '#/components/responses/error_response'

    post:
      summary: 'create a new zone'
      requestBody:
        $ref: '#/components/requestBodies/new_zone'
      responses:
        201:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        409:
          $ref: '#/components/responses/error_response'
  /zones/{zone}:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
    get:
      summary: 'get zone data'
      responses:
        200:
          description: 'successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/zone'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    put:
      summary: 'update zone data'
      requestBody:
        $ref: '#/components/requestBodies/update_zone'
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    delete:
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/import:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
    post:
      summary: 'import zone data'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/export:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
    get:
      summary: 'export zone data'
      responses:
        200:
          description: 'successful response'
          content:
            text/plain:
              schema:
                type: string
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/locations:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string

    get:
      summary: 'list of all locations within zone'
      parameters:
        - $ref: '#/components/parameters/name'
        - $ref: '#/components/parameters/start'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/ascending'
        - $ref: '#/components/parameters/q'
      responses:
        200:
          $ref: '#/components/responses/list_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    post:
      summary: 'create a new location in zone'
      requestBody:
        $ref: '#/components/requestBodies/new_location'
      responses:
        201:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        409:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/locations/{location}:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
      - name: location
        in: path
        required: true
        schema:
          type: string

    get:
      summary: 'get location of a zone'
      responses:
        200:
          description: 'successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/location'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    put:
      summary: 'update a location in zone'
      requestBody:
        $ref: '#/components/requestBodies/update_location'
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    delete:
      summary: 'remove a location from zone'
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/locations/{location}/rrsets:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
      - name: location
        in: path
        required: true
        schema:
          type: string

    get:
      summary: 'list of all rrsets within location'
      responses:
        200:
          $ref: '#/components/responses/list_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    post:
      summary: 'create a new rrset in location'
      requestBody:
        $ref: '#/components/requestBodies/new_rrset'
      responses:
        201:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        409:
          $ref: '#/components/responses/error_response'
  /zones/{zone}/locations/{location}/rrsets/{rtype}:
    parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
      - name: location
        in: path
        required: true
        schema:
          type: string
      - name: rtype
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/rtype'

    get:
      summary: 'get rrset of a zone'
      responses:
        200:
          description: 'successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rrset'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    put:
      summary: 'update a rrset in location'
      requestBody:
        $ref: '#/components/requestBodies/update_rrset'
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'

    delete:
      summary: 'remove a rrset from location'
      responses:
        204:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        404:
          $ref: '#/components/responses/error_response'
...
components:
  parameters:
    start:
      name: start
      in: query
      description: 'start index'
      required: false
      schema:
        type: integer
    count:
      name: count
      in: query
      description: 'number of items in response'
      required: false
      schema:
        type: integer
    ascending:
      name: ascending
      in: query
      description: 'ascending order'
      required: false
      schema:
        type: boolean
    q:
      name: q
      in: query
      description: 'search query'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    new_zone:
      description: 'zone'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/new_zone'
    update_zone:
      description: 'zone'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update_zone'
    new_location:
      description: 'new location'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/new_location'
    update_location:
      description: 'update location'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update_location'
    new_rrset:
      description: new rrset
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/new_rrset'
    update_rrset:
      description: update rrset
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update_rrset'
...
  responses:
    list_response:
      description: 'list response'
      content:
        application/json:
          schema:
            title: list response
            type: object
            properties:
              total:
                type: integer
              items:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    enabled:
                      type: boolean
    error_response:
      description: 'error response'
      content:
        application/json:
          schema:
            title: error response
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
    success_response:
      description: 'successful response'
      content:
        application/json:
          schema:
            title: successful response
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
              "data":
                type: object
  schemas:
...
    new_zone:
      title: new zone
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        dnssec:
          type: boolean
        cname_flattening:
          type: boolean
      example: '{"name": "example.com.", "enabled": true, "dnssec":false, "cname_flattening": false}'
    update_zone:
      title: update zone
      type: object
      properties:
        enabled:
          type: boolean
        dnssec:
          type: boolean
        cname_flattening:
          type: boolean
        soa:
          $ref: '#/components/schemas/soa'
      example: '{"enabled": true, "dnssec":true, "cname_flattening": false, "soa":{"ttl": 300, "ns": "ns1.example.com.", "mbox": "admin.example.com.", "refresh": 44, "retry": 55, "expire": 66, "minttl": 100}}'
    zone:
      title: zone
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        dnssec:
          type: boolean
        cname_flattening:
          type: boolean
        soa:
          $ref: '#/components/schemas/soa'
      example: '{"name": "example.com.", "enabled": true, "dnssec":true, "cname_flattening": false, "soa":{"ttl": 300, "ns": "ns1.example.com.", "mbox": "admin.example.com.", "refresh": 44, "retry": 55, "expire": 66, "minttl": 100}}'
    new_location:
      title: new location
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
      example: '{"name": "www", "enabled": true}'
    update_location:
      title: update location
      type: object
      properties:
        enabled:
          type: boolean
      example: '{"enabled": true}'
    location:
      title: location
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
      example: '{"name": "www", "enabled": true}'
    new_rrset:
      title: new rrset
      type: object
      properties:
        type:
          $ref: '#/components/schemas/rtype'
        value:
          anyOf:
            - $ref: '#/components/schemas/a'
            - $ref: '#/components/schemas/aaaa'
            - $ref: '#/components/schemas/cname'
            - $ref: '#/components/schemas/txt'
            - $ref: '#/components/schemas/ns'
            - $ref: '#/components/schemas/mx'
            - $ref: '#/components/schemas/srv'
            - $ref: '#/components/schemas/caa'
            - $ref: '#/components/schemas/ptr'
            - $ref: '#/components/schemas/tlsa'
            - $ref: '#/components/schemas/ds'
            - $ref: '#/components/schemas/aname'
        enabled:
          type: boolean
...
    rrset:
      title: rrset
      type: object
      properties:
        value:
          anyOf:
            - $ref: '#/components/schemas/a'
            - $ref: '#/components/schemas/aaaa'
            - $ref: '#/components/schemas/cname'
            - $ref: '#/components/schemas/txt'
            - $ref: '#/components/schemas/ns'
            - $ref: '#/components/schemas/mx'
            - $ref: '#/components/schemas/srv'
            - $ref: '#/components/schemas/caa'
            - $ref: '#/components/schemas/ptr'
            - $ref: '#/components/schemas/tlsa'
            - $ref: '#/components/schemas/ds'
            - $ref: '#/components/schemas/aname'
        enabled:
          type: boolean
    rtype:
      type: string
      enum: [a, aaaa, cname, txt, ns, mx srv, caa, ptr, tlsa, ds, aname]
    soa:
      title: soa
      properties:
        ns:
          type: string
        mbox:
          type: string
        refresh:
          type: integer
        retry:
          type: integer
        expire:
          type: integer
        minttl:
          type: integer
        serial:
          type: integer
    a:
      title: a
      allOf:
        - $ref: '#/components/schemas/ip'
    aaaa:
      title: aaaa
      allOf:
        - $ref: '#/components/schemas/ip'
...